Create PROGRAM-2
Converts it to pounds. There are 2.2 pounds in a kilogram. ,Develop a program that uses a for loop to print the numbers 8, 11, 14, 17, 20, . . . , 83, 86, 89.,Develop a program that Split a string into array of characters in Python.,Develop a Python program to get the largest number from a list.,Develop a Python program to calculate the nth Fibonacci number using a function.